Splitting Calendar Factory to its own file
